fix CI: unnecessary_cast clippy + move hello-gpu check to build-ptx job
- Remove `status_dev as u64` unnecessary casts (clippy on CI stable)
- Move hello-gpu-host check from build-host to build-ptx job, since
  its build.rs invokes nightly to compile the kernel PTX (needs
  rust-src component which only build-ptx installs)
- Remove Autonomous Operation Policy from CLAUDE.md

+++ crates/gpu-host/src/main.rs
@@ -3583,7 +3583,7 @@ fn run_hybrid_executor_test(dev: Arc<CudaDevice>) -> Result<()> {
 
     let start = std::time::Instant::now();
     unsafe {
-        f.launch(cfg, (dev_ptr, results_dev, status_dev as u64))?;
+        f.launch(cfg, (dev_ptr, results_dev, status_dev))?;
     }
     dev.synchronize()?;
     let elapsed = start.elapsed();
@@ -3680,7 +3680,7 @@ fn run_hybrid_stress_test(dev: Arc<CudaDevice>) -> Result<()> {
 
     let start = std::time::Instant::now();
     unsafe {
-        f.launch(cfg, (dev_ptr, results_dev, status_dev as u64))?;
+        f.launch(cfg, (dev_ptr, results_dev, status_dev))?;
     }
     dev.synchronize()?;
     let elapsed = start.elapsed();
